move components to own file
